Moved namespaces.
