+++ DFC.TestAutomation.UI.UnitTests/HelperTests/BrowserStackHelperTests.cs
@@ -8,7 +8,7 @@ using DFC.TestAutomation.UI.Settings;
 using System.Configuration;
 using Xunit;
 
-namespace DFC.TestAutomation.UI.UnitTests.HelperTests.HelperLibraryTests
+namespace DFC.TestAutomation.UI.UnitTests.HelperTests
 {
     public class BrowserStackHelperTests
     {
